Test: 좌석 및 구매 테스트 로직에 공통 유틸 적용
